Move auth to lib/, get oauth working again
