+++ SAMPLE-.env
@@ -1,4 +1,4 @@
 # google API credentials - development
-CLIENT_ID=
-CLIENT_SECRET=
-REDIRECT_URL=
+GOOGLE_CLIENT_ID=
+GOOGLE_CLIENT_SECRET=
+GOOGLE_REDIRECT_URL=
