+++ lib/authenticate.js
@@ -1,10 +1,10 @@
+const debug = require('debug')('lib:authenticate')
 const google = require('googleapis')
 const OAuth2 = google.auth.OAuth2
 
-// Credentials
 const CLIENT_ID = process.env.GOOGLE_CLIENT_ID
 const CLIENT_SECRET = process.env.GOOGLE_CLIENT_SECRET
-const REDIRECT_URL = process.env.REDIRECT_URL
+const REDIRECT_URL = process.env.GOOGLE_REDIRECT_URL
 const oauth2Client = new OAuth2(CLIENT_ID, CLIENT_SECRET, REDIRECT_URL)
 
 const scopes = [
@@ -20,20 +20,21 @@ function getAccessToken (res) {
   return res.redirect(url)
 }
 
-// Set client tokens. Redirects to new calendar on completion.
 function setToken (tokenCode, res) {
-  var url = '/error'
   oauth2Client.getToken(tokenCode, function (err, tokens) {
     if (err) {
-      console.log('Error setting client token - failed to authenticate')
-      return res.redirect(url)
+      debug('Error setting client token - failed to authenticate')
+      return res.redirect('/error')
     } else {
       oauth2Client.setCredentials(tokens)
-      url = '/new_bu_calendar'
-      return res.redirect(url)
+      return res.redirect('/new_bu_calendar')
     }
   })
 }
+
+google.options({
+  auth: oauth2Client
+})
 
 module.exports = {
   api: google,
