+++ package.json
@@ -36,6 +36,6 @@
     "events": {
       "restart": "osascript -e 'display notification \"App restarted due to:\n'$FILENAME'\" with title \"nodemon\"'"
     },
-    "ext": "js json"
+    "ext": "js json env"
   }
 }
