+++ routes/clean.js
@@ -1,4 +1,4 @@
-var authenticate = require('./authenticate.js')
+var authenticate = require('../lib/authenticate.js')
 var oauth2Client = authenticate.client
 var calendarAPI = authenticate.api.calendar('v3')
 var newCalendarId = null
